final, all functionality ready

+++ src/main/java/com/javagda25/jsp_invoicemanager/controller/ProductDeleteServlet.java
@@ -1,5 +1,6 @@
 package com.javagda25.jsp_invoicemanager.controller;
 
+import com.javagda25.jsp_invoicemanager.model.Product;
 import com.javagda25.jsp_invoicemanager.service.ProductService;
 
 import javax.servlet.ServletException;
@@ -8,6 +9,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.Optional;
 
 @WebServlet("/product-delete")
 public class ProductDeleteServlet extends HttpServlet {
@@ -16,11 +18,16 @@ public class ProductDeleteServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         Long productId = Long.parseLong(req.getParameter("productId"));
+        Optional<Product> productToEdit = productService.getById(productId);
 
+        if (productToEdit.isPresent()) {
+            Product product = productToEdit.get();
 
-        productService.deleteProduct(productId);
+            Long invoiceId = product.getInvoice().getId();
 
+            productService.deleteProduct(productId);
 
-        resp.sendRedirect("/product-list");
+            resp.sendRedirect("/product-list?invoiceId=" + invoiceId);
+        }
     }
 }
